requirements.txt
Flask==3.1.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
argon2-cffi==23.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==21.2.0
